requirements.txt
pyarrow==22.0.0
annotated-types==0.7.0
anyio==4.11.0
azure-ai-inference==1.0.0b9
azure-core==1.36.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cohere==5.20.0
colorama==0.4.6
distro==1.9.0
fastapi==0.115.0
fastavro==1.12.1
filelock==3.20.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.1.5
idna==3.11
isodate==0.7.2
jiter==0.12.0
joblib==1.4.2
llvmlite==0.45.1
numba==0.62.1
numpy==2.3.5
openai==2.8.1
packaging==25.0
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pyspark==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.5.2
scipy==1.16.3
shap==0.44.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.7
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.31.1
watchfiles==1.1.1
websockets==15.0.1
xgboost==2.1.1